frontend cleanup
